auth/requirements.txt
Django>=5.0
djangorestframework
gunicorn
psycopg2-binary>=2.9
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
requests